feat(pype.lib): restructure .lib to ./lib
